Working on build.
All 5 input configurations are working.  Unit switching is working.  Plot is working, however, entity labels are not being shown on the plot.

TODO: Implement 'settings' functionality to remember settings of units and input configuration.  Review the MyTextBox class, it may not be needed.  Review how the plot is translated in the picture box control.