Requesting pull to master. (#1)
* Establish relationships between project files

When opening the .sln file, the project file was not found in the expected location.  It seems that files were added using the "folder view" of the solution explorer instead of the solution view, which would presumably create the files without linking them to the solution.

* BuildingSolution

Code compiles.  The "AAS" case has been validated.  Scope of work is far greater than the name of the branch implies.  This branch needs to be pulled to the master.  From there, it should be taken back down into a generic "Build" branch.